# backend/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "HTTP API онлайн-сервиса для IT-школы.",
        "title": "skadi API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/user": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение списка юзеров со всеми данными (с настраиваемой пагинацией).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Получение списка юзеров. [Админ и преподаватель]",
                "operationId": "user-list",
                "parameters": [
                    {
                        "type": "boolean",
                        "example": true,
                        "description": "return only class-free students (only if param role=student)",
                        "name": "free",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "example": 1,
                        "description": "page pagination param",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "example": 5,
                        "description": "per page pagination param",
                        "name": "per-page",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "example": [
                            "student"
                        ],
                        "description": "user roles (accepted: \"teacher\", \"student\")",
                        "name": "role",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "иванов",
                        "description": "substring to filter users by username and fullname (case-insensitive)",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.listUserOut"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Создание нового юзера и профиля для него со всеми данными.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Создание нового юзера. [Только админ]",
                "operationId": "user-create",
                "parameters": [
                    {
                        "description": "userBody",
                        "name": "userBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.userBody"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    },
                    "400": {
                        "description": "группа не найдена"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "409": {
                        "description": "пользователь с введенным логином уже существует"
                    }
                }
            }
        },
...
        "/user/{id}": {
            "get": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Получение юзера со всеми данными по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Получение юзера по id. [Только админ]",
                "operationId": "user-read",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID юзера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "404": {
                        "description": "пользователь не найден"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Полное обновление профиля юзера и его группы (если студент) по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Обновление юзера по id. [Только админ]",
                "operationId": "user-update",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID юзера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "updateBody",
                        "name": "updateBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_app_user_controller_http_v1.updateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/entity.User"
                        }
                    },
                    "400": {
                        "description": "неподдерживаемые данные для роли обновляемого пользователя | группа не найдена"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "404": {
                        "description": "пользователь не найден"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Удаление юзера и его профиля по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Удаление юзера по id. [Только админ]",
                "operationId": "user-delete",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID юзера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "404": {
                        "description": "пользователь не найден"
                    }
                }
            }
        },
        "/user/{id}/password": {
            "put": {
                "security": [
                    {
                        "JWTAccess": []
                    }
                ],
                "description": "Смена пароля юзера по его id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Смена пароля юзера. [Только админ]",
                "operationId": "user-password-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID юзера",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "updatePasswordAdminBody",
                        "name": "updatePasswordAdminBody",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1.updatePasswordAdminBody"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "неверный токен (пустой, истекший или неверный формат)"
                    },
                    "404": {
                        "description": "пользователь не найден"
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.Class": {
            "type": "object",
            "required": [
                "id",
                "name"
            ],
            "properties": {
                "id": {
                    "description": "class id",
                    "type": "integer"
                },
                "name": {
                    "description": "class name",
                    "type": "string"
                },
                "schedule": {
                    "description": "class schedule",
                    "type": "string"
                },
                "students": {
                    "description": "student objects list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Profile"
                    }
                },
                "teacher": {
                    "description": "teacher object",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                }
            }
        },
        "entity.Contact": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "entity.Pagination": {
            "type": "object",
            "required": [
                "page",
                "pages",
                "per_page",
                "total"
            ],
            "properties": {
                "page": {
                    "description": "current page",
                    "type": "integer",
                    "example": 1
                },
                "pages": {
                    "description": "total pages amount",
                    "type": "integer",
                    "example": 7
                },
                "per_page": {
                    "description": "per page objects",
                    "type": "integer",
                    "default": 10,
                    "example": 5
                },
                "total": {
                    "description": "total objects amount",
                    "type": "integer",
                    "example": 32
                }
            }
        },
        "entity.Profile": {
            "type": "object",
            "required": [
                "fullname"
            ],
            "properties": {
                "address": {
                    "description": "user address",
                    "type": "string"
                },
                "contact": {
                    "$ref": "#/definitions/entity.Contact"
                },
                "extra": {
                    "description": "extra data (for admin only)",
                    "type": "string"
                },
                "fullname": {
                    "description": "user full name",
                    "type": "string"
                },
                "id": {
                    "description": "profile id (equals to user id)",
                    "type": "integer"
                },
                "parent_contact": {
                    "$ref": "#/definitions/entity.Contact"
                }
            }
        },
...
        "entity.User": {
            "type": "object",
            "required": [
                "id",
                "profile",
                "role",
                "username"
            ],
            "properties": {
                "class": {
                    "description": "student's class",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Class"
                        }
                    ]
                },
                "id": {
                    "description": "user id",
                    "type": "integer"
                },
                "profile": {
                    "description": "user profile",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Profile"
                        }
                    ]
                },
                "role": {
                    "description": "admin, teacher or student",
                    "type": "string"
                },
                "username": {
                    "description": "user username",
                    "type": "string"
                }
            }
        },
...
        "internal_app_user_controller_http_v1.updateBody": {
            "description": "updateBody represents a data to update user and profile.",
            "type": "object",
            "required": [
                "profile"
            ],
            "properties": {
                "class_id": {
                    "description": "class id (for students)",
                    "type": "integer",
                    "example": 3
                },
                "profile": {
                    "description": "user profile",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.profileBody"
                        }
                    ]
                }
            }
        },
...
        "v1.contactBody": {
            "description": "contactBody represents a data with profile contact.",
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "ivanovvp@gmail.com"
                },
                "phone": {
                    "type": "string",
                    "maxLength": 15,
                    "example": "88005553535"
                }
            }
        },
...
        "v1.listUserOut": {
            "description": "listUserOut represents a users list and pagination params.",
            "type": "object",
            "required": [
                "data"
            ],
            "properties": {
                "data": {
                    "description": "users list",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.User"
                    }
                },
                "pagination": {
                    "description": "pagination params",
                    "allOf": [
                        {
                            "$ref": "#/definitions/entity.Pagination"
                        }
                    ]
                }
            }
        },
        "v1.profileBody": {
            "description": "profileBody represents a data with user profile.",
            "type": "object",
            "required": [
                "fullname"
            ],
            "properties": {
                "address": {
                    "description": "user address",
                    "type": "string",
                    "maxLength": 255,
                    "example": "3-я ул.Строителей д. 25"
                },
                "contact": {
                    "description": "user contact info",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.contactBody"
                        }
                    ]
                },
                "extra": {
                    "description": "extra data (for admin only)",
                    "type": "string",
                    "example": "только для админов"
                },
                "fullname": {
                    "description": "user full name",
                    "type": "string",
                    "maxLength": 150,
                    "example": "Иванов Василий Петрович"
                },
                "parent_contact": {
                    "description": "parent contact info (for students)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.contactBody"
                        }
                    ]
                }
            }
        },
...
        "v1.updatePasswordAdminBody": {
            "description": "updatePasswordAdminBody represents a data to update user password by admin.",
            "type": "object",
            "required": [
                "new"
            ],
            "properties": {
                "new": {
                    "description": "new password for user",
                    "type": "string",
                    "maxLength": 40,
                    "minLength": 8,
                    "example": "ytrewq321"
                }
            }
        },
...
        "v1.userBody": {
            "description": "userBody represents a data with user.",
            "type": "object",
            "required": [
                "password",
                "profile",
                "role",
                "username"
            ],
            "properties": {
                "class_id": {
                    "description": "class id (for students)",
                    "type": "integer",
                    "example": 3
                },
                "password": {
                    "description": "user password",
                    "type": "string",
                    "maxLength": 40,
                    "minLength": 8,
                    "example": "qwerty123"
                },
                "profile": {
                    "description": "user profile",
                    "allOf": [
                        {
                            "$ref": "#/definitions/v1.profileBody"
                        }
                    ]
                },
                "role": {
                    "description": "user role (teacher or student)",
                    "type": "string",
                    "enum": [
                        "teacher",
                        "student"
                    ],
                    "example": "teacher"
                },
                "username": {
                    "description": "user username",
                    "type": "string",
                    "maxLength": 50,
                    "example": "user1"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWTAccess": {
            "description": "Access JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "access",
            "in": "cookie"
        },
        "JWTRefresh": {
            "description": "Refresh JWT-token. Cookie will automatic add after auth is done.",
            "type": "apiKey",
            "name": "refresh",
            "in": "cookie"
        }
    }
}